refactor: agent_history/post.py -> agent_history/client.py
